requirements.txt
pandas
liac-arff
scikit-learn
matplotlib
numpy
seaborn